+++ src/Protocol.c
@@ -534,7 +534,7 @@ static void Classic_LevelFinalise(cc_uint8* data) {
 	int delta;
 
 	end   = Stopwatch_Measure();
-	delta = (int)Stopwatch_ElapsedMilliseconds(map_receiveBeg, end);
+	delta = Stopwatch_ElapsedMilliseconds(map_receiveBeg, end);
 	Platform_Log1("map loading took: %i", &delta);
 	map_begunLoading = false;
 	WoM_CheckSendWomID();
